+++ Core/CryptoExchanges/Models/MarketMinimalAggregatedTrade.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 struct MarketMinimalAggregatedTrade {
-    var price: Double
-    var quantity: Double
+    var price: Decimal
+    var quantity: Decimal
     var time: Date
 }
